backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
redis
celery
playwright
beautifulsoup4
psycopg2-binary